requirements.txt
python-telegram-bot==20.7
google-genai
pymupdf
pillow
pytesseract
python-docx